requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.7.3
idna==2.6
pytz==2017.3
requests==2.18.4
retrying==1.3.3
scrapinghub==2.0.1
six==1.11.0
urllib3==1.22
